Got the ambient thing working to the point where I can call it done. Started in on a thing for live performance.
The ambient thing runs off the file temple.rns (Reason file) and temple.rb. It plays nice slow
ambient music until the end of time.

The live performance thing uses yurt.rb, yurt.rns, fuck.rns, and the files in live/. The live thing
can do chord progressions; everything else in this commit can also do key changes, basslines, etc.,
I think that was all in the last commit and since then I've only been tweaking the ambient one to
make it sound better, but I'm not really sure. The past few days have involved a lot of last-minute
logistical panicking.

This process, getting stuff ready, is uncovering loads of flaws. I've got a to-do list going on.
Maybe upgrade to a Lighthouse account near-future.

The chord progressions work pretty well. It's actually almost at a point of sounding good for the
dance music use case when you just leave it running. The goal is this is something I can play live
and tweak. In practice it's not totally there yet. I might be able to hack on it *at* Burning Man
but it seems unlikely, given the cost, the once-a-year-ness, the it-only-it-happens-here-ness, the
general wonder and indescribability, the fact that many of the women are topless, and the fact that
I can hack on my laptop more or less anywhere anytime.

Burning Man is kind of the opposite of Hollywood in that it's a very, very failure-tolerant
environment for artistic experiments. Obviously this only applies to the art side of things; the
desert is not failure-tolerant from a survival point of view at all. But my point is, this live
stuff, I knew it would be good enough for a one-hour live set in a failure-tolerant artistic
environment. I'm beginning to see how it can become good enough for hour-plus live sets in highly
*competitive* artistic environments in the reasonably near future.